- fixed issue #0091 (check for enclosed blocked areas is now sync to game timer) - fixed compile errors in map creator - fixed several smaller bugs

+++ jsettlers.logic.map/jsettlers/logic/map/newGrid/objects/ObjectsGrid.java
@@ -203,7 +203,7 @@ public final class ObjectsGrid implements Serializable {
 		}
 	}
 
-	public Building getBuildingOn(short x, short y) {
+	public Building getBuildingOn(int x, int y) {
 		return buildingsGrid[x + y * width];
 	}
 }
